Add JWT expiration time configuration and refactor authentication classes
