--- src/main/java/oneseoktwojo/ohtalkhae/domain/auth/AuthRepository.java
@@ -1,12 +1,0 @@
-package oneseoktwojo.ohtalkhae.domain.auth;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface AuthRepository extends JpaRepository<User, Long> {
-    Boolean existsByUsername(String username);
-    Boolean existsByEmail(String email);
-    Boolean existsByPhone(String phone);
-    User findByUsername(String username);
-}
