--- src/main/java/oneseoktwojo/ohtalkhae/domain/auth/User.java
@@ -1,36 +1,0 @@
-package oneseoktwojo.ohtalkhae.domain.auth;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import oneseoktwojo.ohtalkhae.domain.auth.enums.Role;
-import org.springframework.data.annotation.CreatedDate;
-
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-
-@Entity
-@Table(name = "users")
-@Getter @Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class User {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long userId;
-    @Column(unique = true)
-    private String username;
-    private String password;
-    @Column(unique = true)
-    private String phone;
-    @Column(unique = true)
-    private String email;
-    private LocalDate birthday;
-    private Long point;
-    @Enumerated(EnumType.STRING)
-    private Role role;
-    @CreatedDate
-    private LocalDateTime createdAt;
-}
