+++ src/main/resources/application.yml
@@ -6,6 +6,8 @@ spring:
       add-mappings: false
   jwt:
     secret: ${JWT_SECRET}
+    refresh-token-expiration-time: 43200000
+    access-token-expiration-time: 3600000
   datasource:
     url: jdbc:h2:mem:ohtalkhae
     driver-class-name: org.h2.Driver
